# dist/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"UserSignUpResponse": {
				"properties": {
					"email": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"preferCategory": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"email",
					"name",
					"preferCategory"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserSignUpRequest": {
				"properties": {
					"email": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"gender": {
						"type": "string"
					},
					"birth": {
						"type": "string"
					},
					"address": {
						"type": "string"
					},
					"detailAddress": {
						"type": "string"
					},
					"phoneNumber": {
						"type": "string"
					},
					"preferences": {
						"items": {
							"type": "number",
							"format": "double"
						},
						"type": "array"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"email",
					"name",
					"gender",
					"birth",
					"phoneNumber",
					"preferences",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateStoreRequest": {
				"properties": {
					"name": {
						"type": "string"
					},
					"address": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"address"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateReviewRequest": {
				"properties": {
					"content": {
						"type": "string"
					},
					"score": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"content",
					"score"
				],
				"type": "object",
				"additionalProperties": false
			},
			"MyChallengingMissionsResponse": {
				"properties": {
					"data": {
						"items": {},
						"type": "array"
					},
					"pagination": {
						"properties": {
							"cursor": {
								"type": "number",
								"format": "double",
								"nullable": true
							}
						},
						"required": [
							"cursor"
						],
						"type": "object"
					}
				},
				"required": [
					"data",
					"pagination"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "umc_10th_node.js",
		"version": "1.0.0",
		"description": "umc프로젝트",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
		"/users/signup": {
			"post": {
				"operationId": "HandleUserSignUp",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserSignUpResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UserSignUpRequest"
							}
						}
					}
				}
			}
		},
		"/users/guest": {
			"get": {
				"operationId": "HandleGuestPage",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": []
			}
		},
		"/users/login": {
			"get": {
				"operationId": "HandleLoginPage",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": []
			}
		},
		"/users/mypage": {
			"get": {
				"operationId": "HandleMypage",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": []
			}
		},
		"/users/set-login": {
			"get": {
				"operationId": "HandleSetLogin",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": []
			}
		},
		"/users/set-logout": {
			"get": {
				"operationId": "HandleSetLogout",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": []
			}
		},
		"/regions/{regionId}/stores": {
			"post": {
				"operationId": "HandleAddStore",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"address": {
											"type": "string"
										},
										"name": {
											"type": "string"
										},
										"regionId": {
											"type": "number",
											"format": "double"
										},
										"storeId": {
											"type": "number",
											"format": "double"
										}
									},
									"required": [
										"address",
										"name",
										"regionId",
										"storeId"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Stores"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "regionId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateStoreRequest"
							}
						}
					}
				}
			}
		},
		"/reviews/me": {
			"get": {
				"operationId": "HandleListMyReviews",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"pagination": {
											"properties": {
												"cursor": {}
											},
											"required": [
												"cursor"
											],
											"type": "object"
										},
										"data": {
											"items": {
												"properties": {
													"score": {},
													"content": {},
													"storeName": {},
													"storeId": {},
													"reviewId": {}
												},
												"required": [
													"score",
													"content",
													"storeName",
													"storeId",
													"reviewId"
												],
												"type": "object"
											},
											"type": "array"
										}
									},
									"required": [
										"pagination",
										"data"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Reviews"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "cursor",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/stores/{storeId}/reviews": {
			"post": {
				"operationId": "HandleAddReview",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"score": {
											"type": "number",
											"format": "double"
										},
										"content": {
											"type": "string"
										},
										"storeId": {
											"type": "number",
											"format": "double"
										},
										"reviewId": {
											"type": "number",
											"format": "double"
										}
									},
									"required": [
										"score",
										"content",
										"storeId",
										"reviewId"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Reviews"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "storeId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateReviewRequest"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "HandleListStoreReviews",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"pagination": {
											"properties": {
												"cursor": {}
											},
											"required": [
												"cursor"
											],
											"type": "object"
										},
										"data": {
											"items": {
												"properties": {
													"store": {},
													"user": {},
													"storeId": {},
													"userId": {},
													"score": {},
													"content": {},
													"id": {}
												},
												"required": [
													"store",
													"user",
													"storeId",
													"userId",
													"score",
													"content",
													"id"
												],
												"type": "object"
											},
											"type": "array"
										}
									},
									"required": [
										"pagination",
										"data"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Reviews"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "storeId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "cursor",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/missions/{missionId}/challenge": {
			"post": {
				"operationId": "HandleChallengeMission",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"status": {
											"type": "string"
										},
										"missionId": {
											"type": "number",
											"format": "double"
										},
										"userMissionId": {
											"type": "number",
											"format": "double"
										}
									},
									"required": [
										"status",
										"missionId",
										"userMissionId"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Missions"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "missionId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/missions/challenging": {
			"get": {
				"operationId": "HandleListMyChallengingMissions",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MyChallengingMissionsResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Missions"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "cursor",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/missions/{missionId}/complete": {
			"patch": {
				"operationId": "HandleCompleteMission",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"status": {
											"type": "string"
										},
										"missionId": {
											"type": "number",
											"format": "double"
										}
									},
									"required": [
										"status",
										"missionId"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Missions"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "missionId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}